requirements (1).txt
```
Python>=3.9
numpy
networkx
matplotlib
streamlit
Pillow
PyPDF2
python-docx
pandas
openpyxl
chromadb
langchain-community
```